world/requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dj-database-url==2.0.0
django-cors-headers==3.14.0
Django==4.2.1
djwto==0.0.4
flake8==6.0.0
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
pathspec==0.11.1
platformdirs==3.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
requests==2.30.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==2.0.2
whitenoise==6.4.0
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.2
